+++ CHANGELOG.md
@@ -4,5 +4,4 @@
 
 **Important note: A completely revised version of the addon will be released in the next few weeks. Keep an eye out for the alpha version and feel free to help me with testing.**
 
-- Added: TOC version for patch 11.2.0 [retail]
-- Deprecated: The screenshot statistics [all versions]
+- Removed: The screenshot statistics [all versions]